+++ tools/local-da/local.go
@@ -19,7 +19,7 @@ import (
 )
 
 // DefaultMaxBlobSize is the default max blob size
-const DefaultMaxBlobSize uint64 = 2 * 1024 * 1024 // 2MB
+const DefaultMaxBlobSize uint64 = 7 * 1024 * 1024 // 7MB
 
 // LocalDA is a simple implementation of in-memory DA. Not production ready! Intended only for testing!
 //